web-nightly-r1/Currency/backend/requirements.txt
# Kconvert - Minimal Production Dependencies
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
slowapi==0.1.9
pydantic==2.9.2
